Made changes to account transfer
Interaccount trasnfer function still allowed user to transfer to a third party account even though it was suppose to check if the receiving account was own by the same customer. Moved the checking logic to the bank method instead and it works. Also joined the both thridparty and inter account transfer method into one